chore: rename lib/common/stdistream.h => lib/common/StdIStream.h
